Adds audio and slides files for how to store crypto lesson
