+++ src/assets/courseData/courses/intro/whyCrypto.js
@@ -1,39 +1,21 @@
-import slide1 from '@/assets/slides/whycrypto/slide1.jpg'
-import slide2 from '@/assets/slides/whycrypto/slide2.jpg'
-import slide3 from '@/assets/slides/whycrypto/slide3.jpg'
-import slide4 from '@/assets/slides/whycrypto/slide4.jpg'
-import slide5 from '@/assets/slides/whycrypto/slide5.jpg'
-import slide6 from '@/assets/slides/whycrypto/slide6.jpg'
-import slide7 from '@/assets/slides/whycrypto/slide7.jpg'
-import slide8 from '@/assets/slides/whycrypto/slide8.jpg'
-import slide9 from '@/assets/slides/whycrypto/slide9.jpg'
-import slide10 from '@/assets/slides/whycrypto/slide10.jpg'
-import slide11 from '@/assets/slides/whycrypto/slide11.jpg'
-import slide12 from '@/assets/slides/whycrypto/slide12.jpg'
-import slide13 from '@/assets/slides/whycrypto/slide13.jpg'
-import slide14 from '@/assets/slides/whycrypto/slide14.jpg'
-import slide15 from '@/assets/slides/whycrypto/slide15.jpg'
-import slide16 from '@/assets/slides/whycrypto/slide16.jpg'
-import slide17 from '@/assets/slides/whycrypto/slide17.jpg'
-import slide18 from '@/assets/slides/whycrypto/slide18.jpg'
-import slide19 from '@/assets/slides/whycrypto/slide19.jpg'
-import slide20 from '@/assets/slides/whycrypto/slide20.jpg'
-import slide21 from '@/assets/slides/whycrypto/slide21.jpg'
-import slide22 from '@/assets/slides/whycrypto/slide22.jpg'
-import slide23 from '@/assets/slides/whycrypto/slide23.jpg'
-import slide24 from '@/assets/slides/whycrypto/slide24.jpg'
-import slide25 from '@/assets/slides/whycrypto/slide25.jpg'
-import slide26 from '@/assets/slides/whycrypto/slide26.jpg'
-import slide27 from '@/assets/slides/whycrypto/slide27.jpg'
-import slide28 from '@/assets/slides/whycrypto/slide28.jpg'
-import slide29 from '@/assets/slides/whycrypto/slide29.jpg'
-import slide30 from '@/assets/slides/whycrypto/slide30.jpg'
-import slide31 from '@/assets/slides/whycrypto/slide31.jpg'
+import slide1 from '@/assets/slides/howtoStore/slide1.jpg'
+import slide2 from '@/assets/slides/howtoStore/slide2.jpg'
+import slide3 from '@/assets/slides/howtoStore/slide3.jpg'
+import slide4 from '@/assets/slides/howtoStore/slide4.jpg'
+import slide5 from '@/assets/slides/howtoStore/slide5.jpg'
+import slide6 from '@/assets/slides/howtoStore/slide6.jpg'
+import slide7 from '@/assets/slides/howtoStore/slide7.jpg'
+import slide8 from '@/assets/slides/howtoStore/slide8.jpg'
+import slide9 from '@/assets/slides/howtoStore/slide9.jpg'
+import slide10 from '@/assets/slides/howtoStore/slide10.jpg'
+import slide11 from '@/assets/slides/howtoStore/slide11.jpg'
+import slide12 from '@/assets/slides/howtoStore/slide12.jpg'
+import slide13 from '@/assets/slides/howtoStore/slide13.jpg'
 import questions from '@/assets/courseData/eduTest/intro/introWhyCryptocurrency.js'
 import notes1 from '@/assets/courseData/courseNotes/intro/whyCrypto/part1.html'
 import notes2 from '@/assets/courseData/courseNotes/intro/whyCrypto/part2.html'
-const why = {
-  title: 'What is Crypto?',
+const howtoStore = {
+  title: 'How to Store Crypto?',
   slides: [
     [
       slide1,
@@ -47,48 +29,23 @@ const why = {
       slide9,
       slide10,
       slide11,
-      slide12,
-      slide13,
-      slide14,
-      slide15
+      slide12
     ],
     [
-      slide16,
-      slide17,
-      slide18,
-      slide19,
-      slide20,
-      slide21,
-      slide22,
-      slide23,
-      slide24,
-      slide25,
-      slide26,
-      slide27,
-      slide28,
-      slide29,
-      slide30
-    ],
-    [
-      slide31
+      slide13
     ]
   ],
   breakpoints: [
     [
-      0, 10, 44, 52, 69, 107, 149, 219,
-      281, 343, 386, 447, 494, 542, 560
-    ],
-    [
-      0, 23, 48, 71, 137, 177, 217, 279, 335, 376, 430, 450, 470, 516, 543
+      0, 30, 57, 104, 144, 170, 258, 369, 489, 611, 649, 706
     ],
     [
       0
     ]
   ],
   audio: [
-    'audio/whatIsCrypto/part1FinalAudio.mp3',
-    'audio/whatIsCrypto/part2Final.mp3',
-    'audio/whatIsCrypto/congratsFinal.mp3'
+    'audio/howtoStore/part1Final.mp3',
+    'audio/howtoStore/congratsFinal.mp3'
   ],
   questions: questions,
   notes: [
@@ -96,4 +53,4 @@ const why = {
     notes2
   ]
 }
-export default why
+export default howtoStore
